Arbitrary framerates, half of level 5
(Higher fps slightly reduces travel distance, likely due to roundoff)

+++ LaunchPad.java
@@ -14,19 +14,19 @@ public class LaunchPad extends MovingObstacle {
         this.launchPower = lPow;
     }
 
-    LaunchPad(int x, int y, int height, int width, XorY XorY, int movePower, int loc1, int loc2, direction dir, int lPow) {
+    LaunchPad(int x, int y, int height, int width, XorY XorY, double movePower, int loc1, int loc2, direction dir, int lPow) {
         super("arrow.png", x, y, height, width, XorY, movePower, loc1, loc2);
         this.dir = dir; 
         this.launchPower = lPow;
     }
 
-    LaunchPad(int x, int y, int height, int width, double xRatio, double yRatio, int movePower, int x1, int x2, int y1, int y2, direction dir, int lPow) {
+    LaunchPad(int x, int y, int height, int width, double xRatio, double yRatio, double movePower, int x1, int x2, int y1, int y2, direction dir, int lPow) {
         super("arrow.png", x, y, height, width, xRatio, yRatio, movePower, x1, x2, y1, y2);
         this.dir = dir; 
         this.launchPower = lPow;
     }
 
-    LaunchPad(int x, int y, int height, int width, boolean moving, double xRatio, double yRatio, int movePower, int x1, int x2, int y1, int y2, direction dir, int lPow) {
+    LaunchPad(int x, int y, int height, int width, boolean moving, double xRatio, double yRatio, double movePower, int x1, int x2, int y1, int y2, direction dir, int lPow) {
         super("arrow.png", x, y, height, width, moving, xRatio, yRatio, movePower, x1, x2, y1, y2);
         this.dir = dir;
         this.launchPower = lPow;
