+++ MovingObstacle.java
@@ -11,7 +11,7 @@ public class MovingObstacle extends MovingSprite {
     }
 
     //x is 0, y is 1
-    MovingObstacle(String imagePath, int x, int y, int height, int width, XorY XorY, int movePower, int loc1, int loc2) {
+    MovingObstacle(String imagePath, int x, int y, int height, int width, XorY XorY, double movePower, int loc1, int loc2) {
         super(imagePath, x, y, height, width);
         moving = true;
         this.movePower = movePower;
@@ -27,7 +27,7 @@ public class MovingObstacle extends MovingSprite {
         }
     }
 
-    MovingObstacle(String imagePath, int x, int y, int height, int width, double xRatio, double yRatio, int movePower, int x1, int x2, int y1, int y2) {
+    MovingObstacle(String imagePath, int x, int y, int height, int width, double xRatio, double yRatio, double movePower, int x1, int x2, int y1, int y2) {
         super(imagePath, x, y, height, width, xRatio, yRatio, movePower);
         this.x1 = x1;
         this.x2 = x2;
@@ -35,7 +35,7 @@ public class MovingObstacle extends MovingSprite {
         this.y2 = y2;
     }
 
-    MovingObstacle(String imagePath, int x, int y, int height, int width, boolean moving, double xRatio, double yRatio, int movePower, int x1, int x2, int y1, int y2) {
+    MovingObstacle(String imagePath, int x, int y, int height, int width, boolean moving, double xRatio, double yRatio, double movePower, int x1, int x2, int y1, int y2) {
         super(imagePath, x, y, height, width, moving, xRatio, yRatio, movePower);
         this.x1 = x1;
         this.x2 = x2;
